--- firmware/stackchan/renderers/manifest.json
@@ -1,22 +1,0 @@
-{
-	"include": [
-		"$(MODDABLE)/examples/manifest_mod.json",
-		"$(MODDABLE)/examples/manifest_typings.json"
-	],
-	"modules": {
-		"*": [
-			"./mod",
-			"../../stackchan/utilities/*"
-		],
-		"renderer-base": "./renderer-base",
-		"effects": "./effects",
-		"simple-face": "./simple-face",
-		"dog-face": "./dog-face"
-	},
-	"preload": [
-		"renderer-base",
-		"simple-face",
-		"dog-face",
-		"effects"
-	]
-}
