+++ firmware/stackchan/renderers/renderer-base.ts
@@ -125,7 +125,7 @@ export function copyFaceContext(src: Readonly<FaceContext>, dst: FaceContext) {
   colorDst[2] = colorSrc[2]
 }
 
-// Modifiers
+// Types
 
 export type FaceModifier<T = unknown> = (tick: number, face: FaceContext, arg?: T) => FaceContext
 export type FaceModifierFactory<T, V = unknown> = (param: T) => FaceModifier<V>
@@ -141,83 +141,6 @@ export type FaceDecorator<T = unknown> = (
   arg?: T
 ) => void
 export type FaceDecoratorFactory<T, V = unknown> = (param: T) => FaceDecorator<V>
-
-function linearInEaseOut(fraction: number): number {
-  if (fraction < 0.25) {
-    return 1 - fraction * 4
-  } else {
-    return (Math.pow(fraction - 0.25, 2) * 16) / 9
-  }
-}
-
-// eslint-disable-next-line @typescript-eslint/no-unused-vars
-function linearInLinearOut(fraction: number): number {
-  if (fraction < 0.5) {
-    return 1 - fraction * 2
-  } else {
-    return fraction * 2 - 1
-  }
-}
-
-export const createBlinkModifier: FaceModifierFactory<{
-  openMin: number
-  openMax: number
-  closeMin: number
-  closeMax: number
-}> = ({ openMin, openMax, closeMin, closeMax }) => {
-  let isBlinking = false
-  let nextToggle = randomBetween(openMin, openMax)
-  let count = 0
-  return (tickMillis: number, face: FaceContext) => {
-    let eyeOpen = 1
-    if (isBlinking) {
-      const fraction = linearInEaseOut(count / nextToggle)
-      eyeOpen = 0.2 + fraction * 0.8
-    }
-    count += tickMillis
-    if (count >= nextToggle) {
-      isBlinking = !isBlinking
-      count = 0
-      nextToggle = isBlinking ? randomBetween(closeMin, closeMax) : randomBetween(openMin, openMax)
-    }
-    Object.values(face.eyes).map((eye) => {
-      eye.open *= eyeOpen
-    })
-    return face
-  }
-}
-
-export const createSaccadeModifier: FaceModifierFactory<{ updateMin: number; updateMax: number; gain: number }> = ({
-  updateMin,
-  updateMax,
-  gain,
-}) => {
-  let nextToggle = randomBetween(updateMin, updateMax)
-  let saccadeX = 0
-  let saccadeY = 0
-  return (tickMillis, face) => {
-    nextToggle -= tickMillis
-    if (nextToggle < 0) {
-      saccadeX = normRand(0, gain)
-      saccadeY = normRand(0, gain)
-      nextToggle = randomBetween(updateMin, updateMax)
-    }
-    Object.values(face.eyes).map((eye) => {
-      eye.gazeX += saccadeX
-      eye.gazeY += saccadeY
-    })
-    return face
-  }
-}
-
-export const createBreathModifier: FaceModifierFactory<{ duration: number }> = ({ duration }) => {
-  let time = 0
-  return (tickMillis, face) => {
-    time += tickMillis % duration
-    face.breath = quantize(Math.sin((2 * Math.PI * time) / duration), 8)
-    return face
-  }
-}
 
 type LayerProps = {
   colorName?: keyof FaceContext['theme']
@@ -271,11 +194,6 @@ export class RendererBase {
     this.layers = []
     this.lastContext = createFaceContext()
     this.currentContext = createFaceContext()
-    this.filters = [
-      createBlinkModifier({ openMin: 400, openMax: 5000, closeMin: 200, closeMax: 400 }),
-      createBreathModifier({ duration: 6000 }),
-      createSaccadeModifier({ updateMin: 300, updateMax: 2000, gain: 0.2 }),
-    ]
     this.clear()
   }
   update(interval = INTERVAL, faceContext: Readonly<FaceContext> = defaultFaceContext): void {
