+++ firmware/stackchan/renderers/simple-face.ts
@@ -1,4 +1,5 @@
 import { RendererBase, Layer, Emotion, type FacePartFactory, type FaceContext } from 'renderer-base'
+import { createBlinkModifier, createBreathModifier, createSaccadeModifier } from 'modifier'
 
 // Renderers
 export const createEyelidPart: FacePartFactory<{
@@ -77,6 +78,11 @@ export const createMouthPart: FacePartFactory<{
 export class Renderer extends RendererBase {
   constructor(option) {
     super(option)
+    this.filters = [
+      createBlinkModifier({ openMin: 400, openMax: 5000, closeMin: 200, closeMax: 400 }),
+      createBreathModifier({ duration: 6000 }),
+      createSaccadeModifier({ updateMin: 300, updateMax: 2000, gain: 0.2 }),
+    ]
     const layer1 = new Layer({ colorName: 'primary' })
     this.layers.push(layer1)
     layer1.addPart(
